wxSlider event handling as described by Vadim in thread 'Slider events' at wx-dev. More changes towards wxRadioBox on Palm.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@31852 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/palmos/control.h
@@ -111,7 +111,7 @@ protected:
                            const wxString& label,
                            const wxPoint& pos,
                            const wxSize& size,
-                           int groupID = 0);
+                           uint8_t groupID = 0);
     inline bool IsPalmControl() const { return m_palmControl; }
 
     bool PalmCreateField(const wxString& label,
